\name{bmd}
\alias{bmd}
\docType{data}
\title{Data of a non-experimental study in more than 300 elderly women}
\description{
Data of a non-experimental study in more than 300 elderly women
}
\usage{data(bmd)}
\format{
  A data frame with 348 observations on the following 5 variables.
  \describe{
    \item{\code{bmd}}{continuous: bmd in mg/cm2}
    \item{\code{age}}{continuous: age in years}
    \item{\code{menopaus}}{continuous: age of menopause}
    \item{\code{weight}}{continuous: weight in kg}
    \item{\code{walkscor}}{dichotomous: score on a walking test, 0=normal, 1=impaired}
  }
}
\examples{
data(bmd)
## maybe str(bmd)
}
\keyword{datasets}






