\name{chol_wide}
\alias{chol_wide}
\docType{data}
\title{Wide dataset of persons from the The Amsterdam Growth and Health Longitudinal Study (AGHLS)}
\description{
Wide dataset of persons from the The Amsterdam Growth and Health Longitudinal Study (AGHLS)
}
\usage{data(chol_wide)}
\format{
  A data frame with 147 observations on the following 7 variables.
  \describe{
    \item{\code{ID}}{continuous}
    \item{\code{Cholesterol1}}{continuous}
    \item{\code{SumSkinfolds1}}{continuous}
    \item{\code{Cholesterol2}}{continuous}
    \item{\code{SumSkinfolds2}}{continuous}
    \item{\code{Cholesterol3}}{continuous}
    \item{\code{SumSkinfolds3}}{continuous}
    \item{\code{Cholesterol4}}{continuous}
    \item{\code{SumSkinfolds4}}{continuous}
    \item{\code{fitness}}{continuous}
    \item{\code{Smoking}}{dichotomous}
    \item{\code{Sex}}{dichotomous}
  }
}
\examples{
data(chol_wide)
## maybe str(chol_wide)
}
\keyword{datasets}
