\name{hoorn_basic}
\alias{hoorn_basic}
\docType{data}
\title{Dataset of the Hoorn Study}
\description{
Dataset of the Hoorn Study
}
\usage{data(hoorn_basic)}
\format{
  A data frame with 250 observations on the following 12 variables.
  \describe{
    \item{\code{patnr}}{continuous}
    \item{\code{sbldsys1}}{continuous: Systolic Blood Pressure 1}
    \item{\code{sbldsys2}}{continuous: Systolic Blood Pressure 2}
    \item{\code{sbldds1}}{continuous: Diastolic Blood Pressure 1}
    \item{\code{sbldds2}}{continuous: Diastolic Blood Pressure 2}
    \item{\code{sex}}{dichotomous: 1=male, 2=female}
    \item{\code{sfructo}}{continuous: fructosamine level in the blood }
    \item{\code{sglucn}}{continuous}
    \item{\code{dmknown}}{dichotomous: 0=no, 1=yes}
    \item{\code{dmdiet}}{dichotomous: 0=no, 1=yes}
    \item{\code{infarct}}{dichotomous: 0=no, 1=yes}
    \item{\code{hypten}}{dichotomous: 0=no, 1=yes}
    
  }
}
\examples{
data(hoorn_basic)
## maybe str(hoorn_basic)
}
\keyword{datasets}
