% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_fit.R
\name{km_fit}
\alias{km_fit}
\title{Kaplan-Meier (KM) estimate at specific time point}
\usage{
km_fit(time, status, t_risk)
}
\arguments{
\item{time}{Character vector. Name of time variable.}

\item{status}{Character vector. Name of status variable.}

\item{t_risk}{Follow-up value to calculate cases, controls. See details.}
}
\value{
KM estimate at specific time point
}
\description{
Kaplan-Meier (KM) estimate at specific time point
}
\references{
Pencina MJ, D'Agostino RB Sr, Steyerberg EW. Extensions of net reclassification 
 improvement calculations to measure usefulness of new biomarkers. Stat Med. 2011;30(1):11-21

Inoue E (2018). nricens: NRI for Risk Prediction Models with Time to Event and Binary 
 Response Data. R package version 1.6, <https://CRAN.R-project.org/package=nricens>.
}
\seealso{
\code{\link{km_fit}}
}
\author{
Martijn Heymans, 2023
}
