\name{lbp_orig}
\alias{lbp_orig}
\docType{data}
\title{Example dataset for psfmi_perform function, method boot_MI}
\description{
Original dataset with missing values
}
\usage{data(lbp_orig)}
\format{
  A data frame with 159 observations on the following 15 variables.
  \describe{
    \item{\code{Chronic}}{dichotomous}
    \item{\code{Gender}}{dichotomous}
    \item{\code{Carrying}}{categorical}
    \item{\code{Pain}}{continuous}
    \item{\code{Tampascale}}{continuous}
    \item{\code{Function}}{continuous}
    \item{\code{Radiation}}{dichotomous}
    \item{\code{Age}}{continuous}
    \item{\code{Smoking}}{dichotomous}
    \item{\code{Satisfaction}}{categorical}
    \item{\code{JobControl}}{continuous}
    \item{\code{JobDemands}}{continuous}
    \item{\code{SocialSupport}}{continuous}
    \item{\code{Duration}}{continuous}
    \item{\code{BMI}}{continuous}
  }
}
\examples{
data(lbp_orig)
## maybe str(lbp_orig)
}
\keyword{datasets}
