\name{lbpmilr_dev}
\alias{lbpmilr_dev}
\docType{data}
\title{Example dataset for mivalext_lr function}
\description{
1 development dataset
}
\usage{data(lbpmilr_dev)}
\format{
  A data frame with 108 observations on the following 16 variables.
  \describe{
    \item{\code{ID}}{a numeric vector}
    \item{\code{Chronic}}{dichotomous}
    \item{\code{Gender}}{dichotomous}
    \item{\code{Carrying}}{categorical}
    \item{\code{Pain}}{continuous}
    \item{\code{Tampascale}}{continuous}
    \item{\code{Function}}{continuous}
    \item{\code{Radiation}}{dichotomous}
    \item{\code{Age}}{continuous}
    \item{\code{Smoking}}{dichotomous}
    \item{\code{Satisfaction}}{categorical}
    \item{\code{JobControl}}{continuous}
    \item{\code{JobDemands}}{continuous}
    \item{\code{SocialSupport}}{continuous}
    \item{\code{Duration}}{continuous}
    \item{\code{BMI}}{continuous}
  }
}
\examples{
data(lbpmilr_dev)
## maybe str(lbpmilr_dev)
}
\keyword{datasets}
