\name{lungvolume}
\alias{lungvolume}
\docType{data}
\title{Data of the development of lung and heartvolume of unborn babies}
\description{
Data regarding the development of lung and heartvolume of unborn babies in the 18e till 34e week of pregnancy
}
\usage{data(lungvolume)}
\format{
  A data frame with 152 observations on the following 6 variables.
  \describe{
    \item{\code{pat_id}}{continuous}
    \item{\code{week}}{continuous: week pregnancy}
    \item{\code{weight}}{continuous: weight in grams}
    \item{\code{lungvol}}{continuous: lung volume}
    \item{\code{heartvol}}{continuous: heart volume}
    \item{\code{Nweek}}{categorical: Percentile Group of week}
  }
}
\examples{
data(lungvolume)
## maybe str(lungvolume)
}
\keyword{datasets}




