% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_auc_log.R
\name{mean_auc_log}
\alias{mean_auc_log}
\title{Function to calulate mean auc values}
\usage{
mean_auc_log(auc, backtransform = TRUE)
}
\description{
Function to calulate mean auc values
}
\author{
Martijn Heymans, 2020
}
\keyword{internal}
