% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI_boot.R
\name{MI_boot}
\alias{MI_boot}
\title{Bootstrap validation in Multiply Imputed datasets}
\usage{
MI_boot(pobj, p.crit, nboot, direction)
}
\arguments{
\item{pobj}{An object of class \code{pmods} (pooled models), produced by a previous
call to \code{psfmi_lr}.}

\item{p.crit}{A numerical scalar. P-value selection criterium used for backward
or forward selection during in the bootstrap samples. When set at 1, validation is done
without variable selection.}

\item{nboot}{The number of bootstrap resamples, default is 10.}

\item{direction}{The direction of predictor selection, "BW" is for backward selection and
"FW" for forward selection.}
}
\description{
\code{MI_boot} Bootstrapping in each (original) Multiply Imputed dataset for internal validation.
 Called by function \code{psfmi_perform}.
}
\seealso{
\code{\link{psfmi_perform}}
}
\author{
Martijn Heymans, 2020
}
\keyword{internal}
