% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_performance_internal.R
\name{pool_performance_internal}
\alias{pool_performance_internal}
\title{Pooling performance measures over multiply imputed datasets}
\usage{
pool_performance_internal(data, formula, nimp, impvar)
}
\arguments{
\item{data}{Data frame with stacked multiple imputed datasets.
The original dataset that contains missing values must be excluded 
from the dataset.}

\item{formula}{A formula object to specify the model as normally used by glm.}

\item{nimp}{A numerical scalar. Number of imputed datasets. Default is 5.}

\item{impvar}{A character vector. Name of the variable that distinguishes 
the imputed datasets.}
}
\description{
\code{pool_performance_internal} Pooling performance measures
}
\keyword{internal}
