% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_reclassification.R
\name{pool_reclassification}
\alias{pool_reclassification}
\title{Function to pool NRI measures over Multiply Imputed datasets}
\usage{
pool_reclassification(datasets, cutoff = cutoff)
}
\arguments{
\item{datasets}{a list of data frames corresponding to the multiply imputed
datasets, within each dataset in the first column the predicted probabilities
of model 1, in the second column those of model 2 and in the third column
the observed outcomes coded as '0'and '1'.}

\item{cutoff}{cutoff value for the categorical NRI, must lie between 0 and 1.}
}
\description{
\code{pool_reclassification} Function to pool categorical and continuous NRI
  and IDI over Multiply Imputed datasets
}
\details{
This function is called by the function \code{pool_compare_model}
}
\author{
Martijn Heymans, 2020
}
