% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_RR.R
\name{pool_RR}
\alias{pool_RR}
\title{Function to combine estimates by using Rubin's Rules}
\usage{
pool_RR(est, se, conf.level = 0.95, n, k)
}
\arguments{
\item{est}{A vector of multiple parameter estimates}

\item{se}{A vector of multiple standard error estimates}

\item{conf.level}{desired confidence limits}

\item{n}{sample size in completed dataset}

\item{k}{number of parameters to pool}
}
\description{
\code{pool_RR} Rubin's Rules
}
\author{
Martijn Heymans, 2021
}
