% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psfmi_lr.R
\name{psfmi_lr}
\alias{psfmi_lr}
\title{Pooling and Predictor selection function for Logistic regression
models in multiply imputed datasets}
\usage{
psfmi_lr(data, nimp = 5, impvar = NULL, Outcome, predictors = NULL,
  p.crit = 1, cat.predictors = NULL, spline.predictors = NULL,
  int.predictors = NULL, keep.predictors = NULL, knots = NULL,
  method = NULL, print.method = FALSE)
}
\arguments{
\item{data}{Data frame or data matrix with stacked multiple imputed datasets.
The original dataset that contains missing values must be excluded from the
dataset. The imputed datasets must be distinguished by an imputation variable,
specified under impvar, and starting by 1.}

\item{nimp}{A numerical scalar. Number of imputed datasets. Default is 5.}

\item{impvar}{A character vector. Name of the variable that distinguishes the
imputed datasets.}

\item{Outcome}{Character vector containing the name of the outcome variable.}

\item{predictors}{Character vector with the names of the predictor variables.
At least one predictor variable has to be defined.}

\item{p.crit}{A numerical scalar. P-value selection criterium.}

\item{cat.predictors}{A single string or a vector of strings to define the
categorical variables. Default is NULL categorical predictors.}

\item{spline.predictors}{A single string or a vector of strings to define the
(restricted cubic) spline variables. Default is NULL spline predictors. See details.}

\item{int.predictors}{A single string or a vector of strings with the names of the variables that form
an interaction pair, separated by a “:” symbol.}

\item{keep.predictors}{A single string or a vector of strings including the variables that are forced
in the model during predictor selection. Categorical and interaction variables are allowed.}

\item{knots}{A numerical vector that defines the number of knots for each spline predictor separately.}

\item{method}{A character vector to indicate the pooling method for p-values to pool the
total model or used during predictor selection. This can be "D1", "D2", "MR" or "MPR".
See details for more information.}

\item{print.method}{logical vector. If TRUE full matrix with p-values of all variables according to
chosen method (under method) is shown. If FALSE (default) p-value for categorical variables according
to method are shown and for continuous and dichotomous predictors Rubin’s Rules are used.}
}
\description{
\code{psfmi_lr} Pooling and backward selection for Logistic regression
prediction models in multiply imputed datasets using different selection methods.
}
\details{
The basic pooling procedure to derive pooled coefficients, standard errors, 95
 confidence intervals and p-values is Rubin's Rules (RR). Specific procedures are
 available to derive pooled p-values for categorical (> 2 categories) and spline variables.
 print.method allows to choose between these pooling methods that are:
 “D1” is pooling of the total covariance matrix, ”D2” is pooling of Chi-square values,
 “MR” is pooling Likelihood ratio statistics (method of Meng and Rubin) and “MPR”
 is pooling of median p-values (MPR rule). Spline regression coefficients are defined
 by using the rcs function for restricted cubic splines of the rms package of Frank Harrell.
 A minimum number of 3 knots as defined under knots is needed.
}
\examples{
  # Pooling model (without backward selection) using D1
  psfmi_lr(data=lbpmilr, nimp=5, impvar="Impnr", Outcome="Chronic",
  predictors=c("Gender", "Smoking", "Function", "JobControl",
  "JobDemands", "SocialSupport"), method="D1")

  # Predictor selection using p<0.05 and method D1
  psfmi_lr(data=lbpmilr, nimp=5, impvar="Impnr", Outcome="Chronic",
  predictors=c("Gender", "Smoking", "Function", "JobControl",
  "JobDemands", "SocialSupport"), p.crit = 0.05, method="D1")

  # Predictor selection using p<0.05 and method MPR
  psfmi_lr(data=lbpmilr, nimp=5, impvar="Impnr", Outcome="Chronic",
  predictors=c("Gender", "Smoking", "Function", "JobControl",
  "JobDemands", "SocialSupport"), p.crit = 0.05, method="MPR")

  # Predictor selection, force variable Smoking in model,
  # using p<0.05 and method MR
  psfmi_lr(data=lbpmilr, nimp=5, impvar="Impnr", Outcome="Chronic",
  predictors=c("Gender", "Smoking", "Function", "JobControl",
  "JobDemands", "SocialSupport"), keep.predictors = "Smoking",
  p.crit = 0.05, method="MR")

  # Predictor selection, including categorical and interaction terms
  # and forcing interaction terms and another variable in model,
  # using p<0.05 and method D1
  psfmi_lr(data=lbpmilr, nimp=5, impvar="Impnr", Outcome="Chronic",
  predictors=c("Gender", "Smoking", "Function", "JobControl", "JobDemands",
  "SocialSupport"), p.crit = 0.05, cat.predictors = c("Carrying", "Satisfaction"),
  int.predictors = c("Carrying:Smoking", "Gender:Smoking"),
  keep.predictors = c("Smoking:Carrying", "JobControl"), method="D1")

  # Predictor selection, including categorical, spline (3 knots) and
  # interaction terms and forcing variable in model, using p<0.05 and method D1
  psfmi_lr(data=lbpmilr, nimp=5, impvar="Impnr", Outcome="Chronic",
  predictors=c("Gender", "Smoking", "JobControl", "JobDemands",
  "SocialSupport"), p.crit = 0.05, cat.predictors = c("Carrying", "Satisfaction"),
  spline.predictors=c("Function"), int.predictors = c("Gender:Smoking"),
  keep.predictors = c("JobControl"), knots=3, method="D1")

}
\references{
Eekhout I, van de Wiel MA, Heymans MW. Methods for significance testing of categorical
  covariates in logistic regression models after multiple imputation: power and applicability
  analysis. BMC Med Res Methodol. 2017;17(1):129.

Enders CK (2010). Applied missing data analysis. New York: The Guilford Press.

Meng X-L, Rubin DB. Performing likelihood ratio tests with multiply-imputed data sets.
  Biometrika.1992;79:103-11.

van de Wiel MA, Berkhof J, van Wieringen WN. Testing the prediction error difference between
  2 predictors. Biostatistics. 2009;10:550-60.

Marshall A, Altman DG, Holder RL, Royston P. Combining estimates of interest in prognostic
  modelling studies after multiple imputation: current practice and guidelines. BMC Med Res Methodol.
  2009;9:57.

Van Buuren S. (2012). Flexible Imputation of Missing Data. Chapman & Hall/CRC
  Interdisciplinary Statistics. Boca Raton.
}
