% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psfmi_lr_fw.R
\name{psfmi_lr_fw}
\alias{psfmi_lr_fw}
\title{Forward selection of Logistic regression models in multiply imputed data.}
\usage{
psfmi_lr_fw(data, nimp, impvar, Outcome, P, p.crit, method, keep.P)
}
\arguments{
\item{data}{Data frame with stacked multiple imputed datasets.
The original dataset that contains missing values must be excluded from the
dataset. The imputed datasets must be distinguished by an imputation variable,
specified under impvar, and starting by 1.}

\item{nimp}{A numerical scalar. Number of imputed datasets. Default is 5.}

\item{impvar}{A character vector. Name of the variable that distinguishes the
imputed datasets.}

\item{Outcome}{Character vector containing the name of the outcome variable.}

\item{P}{Character vector with the names of the predictor variables.
At least one predictor variable has to be defined. Give predictors unique names
and do not use predictor name combinations with numbers as, age2, BMI10, etc.}

\item{p.crit}{A numerical scalar. P-value selection criterium. A value of 1
provides the pooled model without selection.}

\item{method}{A character vector to indicate the pooling method for p-values to pool the
total model or used during predictor selection. This can be "RR", D1", "D2", "D3" or "MPR".
See details for more information. Default is "RR".}

\item{keep.P}{A single string or a vector of strings including the variables that are forced
in the model during predictor selection. Categorical and interaction variables are allowed.}
}
\description{
\code{psfmi_lr_fw} Forward selection of Logistic regression
models across multiply imputed data using selection methods RR, D1, D2, D3 and MPR.
Function is called by \code{psfmi_lr}.
}
\author{
Martijn Heymans, 2020
}
\keyword{internal}
