% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_coxph.R
\name{risk_coxph}
\alias{risk_coxph}
\title{Risk calculation at specific time point for Cox model}
\usage{
risk_coxph(mod, t_risk)
}
\arguments{
\item{mod}{a Cox regression model object.}

\item{t_risk}{Follow-up value to calculate cases, controls. See details.}
}
\value{
Cox regression Risk estimates at specific time point.
}
\description{
Risk calculation at specific time point for Cox model
}
\references{
Pencina MJ, D'Agostino RB Sr, Steyerberg EW. Extensions of net reclassification 
 improvement calculations to measure usefulness of new biomarkers. Stat Med. 2011;30(1):11-21

Inoue E (2018). nricens: NRI for Risk Prediction Models with Time to Event and Binary 
 Response Data. R package version 1.6, <https://CRAN.R-project.org/package=nricens>.
}
\seealso{
\code{\link{nri_cox}}
}
\author{
Martijn Heymans, 2023
}
