% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR_diff_prop.R
\name{RR_diff_prop}
\alias{RR_diff_prop}
\title{Function to apply RR to pool difference of NRI and AUC values}
\usage{
RR_diff_prop(est, se)
}
\description{
Function to apply RR to pool difference of NRI and AUC values
}
\author{
Martijn Heymans, 2020
}
\keyword{internal}
