% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsq_nagel.R
\name{rsq_nagel}
\alias{rsq_nagel}
\title{Nagelkerke's R-square calculation for logistic regression / glm models}
\usage{
rsq_nagel(fitobj)
}
\arguments{
\item{fitobj}{a logistic regression model object of "glm"}
}
\value{
The value for the explained variance.
}
\description{
Nagelkerke's R-square calculation for logistic regression / glm models
}
\seealso{
\code{\link{psfmi_perform}}, \code{\link{pool_performance}}
}
\author{
Martijn Heymans, 2020
}
