% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsq_surv.R
\name{rsq_surv}
\alias{rsq_surv}
\title{R-square calculation for Cox regression models}
\usage{
rsq_surv(fitobj)
}
\arguments{
\item{fitobj}{a Cox regression model object of "coxph"}
}
\value{
The value for the explained variance.
}
\description{
R-square calculation for Cox regression models
}
\references{
F. Harrell. Regression Modeling Strategies. With Applications to
 Linear Models, Logistic and Ordinal Regression, and Survival Analysis. 2nd Edition.
 Springer, New York, NY, 2015.
}
\seealso{
\code{\link{pool_performance}}
}
\author{
Martijn Heymans, 2021
}
