% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled_brier.R
\name{scaled_brier}
\alias{scaled_brier}
\title{Calculated the scaled Brier score}
\usage{
scaled_brier(obs, pred)
}
\arguments{
\item{obs}{Observed outcomes.}

\item{pred}{Predicted outcomes in the form of probabilities.}
}
\value{
The value for the scaled Brier score.
}
\description{
Calculated the scaled Brier score
}
\seealso{
\code{\link{psfmi_perform}}, \code{\link{pool_performance}}
}
\author{
Martijn Heymans, 2020
}
