\name{smoking}
\alias{smoking}
\docType{data}
\title{Survival data about smoking}
\description{
Survival data about smoking
}
\usage{data(smoking)}
\format{
  A data frame with 20 observations on the following 3 variables.
  \describe{
    \item{\code{smoking}}{dichotomous: 1=yes, 0=no}
    \item{\code{time}}{continuous: Survival time in years}
    \item{\code{death}}{dichotomous: Status at end of study}
  }
}
\examples{
data(smoking)
## maybe str(smoking)
}
\keyword{datasets}

