\name{weight}
\alias{weight}
\docType{data}
\title{Dataset of persons from the The Amsterdam Growth and Health Longitudinal Study (AGHLS)}
\description{
Dataset of persons from the The Amsterdam Growth and Health Longitudinal Study (AGHLS)
}
\usage{data(weight)}
\format{
  A data frame with 450 observations on the following 7 variables.
  \describe{
    \item{\code{ID}}{continuous}
    \item{\code{SBP}}{continuous: Systolic Blood Pressure}
    \item{\code{LDL}}{continuous: Cholesterol}
    \item{\code{Glucose}}{continuous}
    \item{\code{HDL}}{continuous: Cholesterol}
    \item{\code{Gender}}{dichotomous: 1=male, 0=female}
    \item{\code{Weight}}{continuous: bodyweight}
  }
}
\examples{
data(weight)
## maybe str(weight)
}
\keyword{datasets}
