\name{sbp_age}
\alias{sbp_age}
\docType{data}
\title{Dataset with blood pressure measurements}
\description{
Dataset with blood pressure measurements
}
\usage{data(sbp_age)}
\format{
  A data frame with 30 observations on the following 3 variables.
  \describe{
    \item{\code{pat_id}}{continuous}
    \item{\code{sbp}}{continuous: systo¬lic blood pressure (SBP, in mm Hg)}
    \item{\code{age}}{continuous: age (AGE, in years)}
  }
}
\examples{
data(sbp_age)
## maybe str(sbp_age)
}
\keyword{datasets}
